Reorder services in order of startup
If PostgreSQL service is started after Adminer, Adminer will not
recognized postgres.

Error: "postgres could not connect to server. Connection refused..."

With the reorder this seems to disappear. Docker notes that services
do not wait for one another to load and there is no nice way to
resolve this.

Should it become an issue then it will be necessary to create a
script that holds off starting adminer until after postgres has
completed. For now, unnecessary.